refactor: 조합 코드 리팩토링
